리팩토링 폴더 분류
