+++ src/components/Header.tsx
@@ -1,11 +1,11 @@
 import { memo, useCallback } from "../@lib";
-import { useThemeContext, useUserContext } from "../contexts";
+import { useAuthContext, useThemeContext } from "../contexts";
 import { renderLog } from "../utils";
 
 export const Header: React.FC = memo(() => {
   renderLog("Header rendered");
   const { theme, toggleTheme } = useThemeContext();
-  const { user, login, logout } = useUserContext();
+  const { user, login, logout } = useAuthContext();
 
   const handleLogin = useCallback(() => {
     // 실제 애플리케이션에서는 사용자 입력을 받아야 합니다.
