+++ src/hooks/index.ts
@@ -1,5 +1,1 @@
-export * from "./useTheme";
 export * from "./useItmes";
-export * from "./useUser";
-export * from "./useNotification";
-export * from "./useAuth";
