--- src/hooks/useTheme.ts
@@ -1,12 +1,0 @@
-import { useState } from "react";
-import { useCallback } from "../@lib";
-
-export const useTheme = () => {
-  const [theme, setTheme] = useState("light");
-
-  const toggleTheme = useCallback(() => {
-    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
-  }, []);
-
-  return { theme, toggleTheme };
-};
